style added

+++ app/employee/list/page.jsx
@@ -23,11 +23,10 @@ const page = () => {
   }, []);
 
   const handleDelete = (id) => {
-    dispatch(deleteEmployees(id, () => dispatch(getEmployees())));
-    // const isConfirmed = window.confirm("Are you sure you want to delete?");
-    // if (isConfirmed) {
-    //   dispatch(deleteEmployees(id, () => dispatch(getEmployees())));
-    // }
+    const isConfirmed = window.confirm("Are you sure you want to delete?");
+    if (isConfirmed) {
+      dispatch(deleteEmployees(id, () => dispatch(getEmployees())));
+    }
   };
 
   const handleEdit = (id) => router.push(`/employee/edit?id=${id}`);
@@ -54,7 +53,7 @@ const page = () => {
       <div
         className={
           gridView
-            ? " px-20 grid lg:grid-cols-5 md:grid-cols-4 grid-cols-3 gap-8 py-4 items-center justify-center align-middle"
+            ? "grid lg:grid-cols-5 md:grid-cols-4 grid-cols-3 px-20 gap-8 py-4 items-center justify-center align-middle"
             : "flex py-10 px-40"
         }
       >
